# Optimise BlockSize::from_width_and_height_opt